+++ gframe/single_mode.h
@@ -35,7 +35,7 @@ public:
 	
 	static void SinglePlayRefresh(uint8_t player, uint8_t location, uint32_t flag = 0x2f81fff);
 	static void SinglePlayRefresh(uint32_t flag = 0x2f81fff);
-	static void SinglePlayRefreshSingle(uint8_t player, uint8_t location, uint32_t sequence, uint32_t flag = 0x2f81fff);
+	static void SinglePlayRefreshSingle(uint8_t player, uint8_t location, uint8_t sequence, uint32_t flag = 0x2f81fff);
 	static void SinglePlayReload();
 	static Signal singleSignal;
